requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
